318 - Generate smooth hydrograph using gamma function for Ewater source data exports and Flood operations Implemented Gamma Function (daily and hourly timestep) in PrepareewSourceRO_Qcomb. This implementation needs further testing - specifically related to the units of the provided values: Phi_Alpha = (Qp * tp) / (Cv * A)
In this formula,
Qp - the peak flow rate Qp (Gamma_peak_value) needs to be converted to units
of "cfs /inch" (this is the peak value of the unit hydrograph in cfs for a
1-inch rainfall depth).
A - drainage area A needs to be defined in acres
tp - time to peak tp in hours
Cv  - equals 1.008 for volumes measured in cfs * hrs, drainage area in acres,
and the unit storm as 1-inch.

It appears the provided value in the attributes are not representative - it is
unclear how the conversion should take place of the provided Qp value.
As it stands, with just a direct conversion from ML/day to cfs, the resulting
value of Phi_Alpha is not of the right order, messing up the lookup for alpha.