R_314: Create indicator timeseries for each location/parameter combination automatically on a daily basis with a default value "1".
?             H
?             dH
?             Q
?             dQ
?             V
?             DO%
Note: specials have not (yet) been included.

In total there are 35 locations that have one or more of the selection of
thresholds defined.

The transformation is set up to not overwrite any existing edits
Note: <parameterId>Threshold_H_exclusion</parameterId> is of the type
?enumeration? ? in order to force only two possible options: 0 or 1.
The user should be able to edit the indicator timeseries and set it to "0"
for a selected period.
0 = NOT counted
1 = counted

Indicator timeseries is integrated in the transformation counting the
threshold crossings. If the indicator time series is 1, a breach at that
time step is counted. If it is 0, any breach in that period will be excluded
from the count

This happens in AssessThresholdsMDBC_OPS_SO.xml
Basically the *_threshold_4count timeseries are computed as regular, but stored
as a temporary timeseries. We then apply a function to exclude crossing values
based on the Threshold_*_exclusion timeseries

Create "Review Threshold" node similar to ?review telemetry? workflow node for
each location that displays the timeseries and the indicator timeseries for the
associated parameters.