R314 Renamed threshold "counting" timeseries. Also added a qualifier to show that the implementation is set up for SO data/thresholds
Explanation of included functionality added to the ModuleConfig file:
The SO "counted" timeseries (generated in PreprocessThresholdsMDBC_OPS) for all
SO Threshold_sites and parameter combinations are used in this module:
If these timeseries are set to 0 by the operator, any threshold violations for that
    location/parameter combination will not be counted.
    The following parameters are supported:
    - H
    - dH
    - Q
    - dQ
    - V
    - DO%
The implementation is set up for the 8am (daily) threshold violations.
dH threshold have some specials (not just daily, also     mov_avg, hourly, etc).
These are currently not affected.
The implementation (in general) takes the results of the *_4count 'merge'
transformation (that combines all the threshold violation timeseries per
parameter) and sets the value to 0 if the  SO "counted" timeseries are 0.
Note1: The thresholds violations themselves are not removed, they will just not be counted.
Note2: Thuis is implemented for the SO threshold crossings, not for the RO